feat: migrate illustrated-message component to TS, small refactor
